pkgconfig-package {pkgconfig}	R Documentation

Persistent configuration for R packages

Description

This package is meant to be used in other packages, and provides configuration options for them.

Details

It is currently very minimal, and has two main functions: one for setting configuration options (set_config), and one for querying them (get_config).

[Package pkgconfig version 2.0.3 Index]