set_config {pkgconfig}	R Documentation

Set a configuration parameter

Description

Set a configuration parameter, for the package we are calling from. If called from the R prompt and not from a package, then it sets the parameter for global environment.

Usage


set_config(...)

Arguments

...
Parameters to set, they should be all named.

Value

Nothing.

See Also

set_config_in

[Package pkgconfig version 2.0.3 Index]