set_config_in {pkgconfig}	R Documentation

Set a configuration parameter for a package

Description

This is a more flexible variant of link{set_config}, and it allows creating an custom API in the package that uses pkgconfig for its configuration.

Usage


set_config_in(..., .in = parent.frame())

Arguments

...
Parameters to set, they should be all named.

.in
An environment, typically belonging to a package.

Details

This function is identical to set_config, but it allows supplying the environment that is used as the package the configuration is set for. This makes it possible to create an API for setting (and getting) configuration parameters.

Value

Nothing.

See Also

set_config

[Package pkgconfig version 2.0.3 Index]